backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
yarg==0.1.9
